libbpf: add new strict flag for .text subprograms
Currently, libbpf considers a single routine in .text as a program. This
is particularly confusing when it comes to library objects - a single routine
meant to be used as an extern will instead be considered a bpf_program.

This patch hides this compatibility behavior behind a libbpf_mode strict
mode flag.

Signed-off-by: Delyan Kratunov <delyank@fb.com>